token logo parsing from jupiter list
