+++ packages/data-components/src/utils.ts
@@ -23,3 +23,22 @@ export function formatDate(date: Date): string {
   const yyyy = date.getFullYear();
   return `${mm} ${dd}, ${yyyy}`;
 }
+
+/**
+ * Convert a snake case string into normalized title case.
+ * @export
+ * @param {string} input
+ * @returns {string}
+ */
+export function snakeToTitleCase(input: string): string {
+  const parts = input.split("_").map((t) => t.toLowerCase());
+  const titleCasesParts = parts.map((p) =>
+    p.length === 1 ? p : `${p[0].toUpperCase()}${p.slice(1)}`
+  );
+
+  if (titleCasesParts[0] === "Nft") {
+    titleCasesParts[0] = titleCasesParts[0].toUpperCase();
+  }
+
+  return titleCasesParts.join(" ");
+}
